requirements.txt

flaskblinker==1.9.0
cffi==2.0.0
click==8.3.2
cryptography==41.0.7
Flask==3.1.3
grpcio==1.80.0
grpcio-tools==1.80.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
protobuf==6.33.6
pycparser==3.0
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2026.1.post1
PyYAML==6.0.3
setuptools==82.0.1
typing_extensions==4.15.0
Werkzeug==3.1.8
